Organize code
This commit moves the code from the React Native demo app into
its own component.
Source files will be under `src/`.
I made different entry points:
index.js -> App.tsx -> main.tsx -> MainAppCtl
