+++ src/MainApp/MainAppCtl.tsx
@@ -1,0 +1,11 @@
+/**
+ * This is the main app component
+ * This is where core setup and providers will be handled
+ */
+
+import React from 'react';
+import RNDemoApp from '../RNDemoApp/RNDemoApp';
+
+export const MainAppCtl = () => {
+  return <RNDemoApp />;
+};
