+++ src/main.tsx
@@ -1,0 +1,10 @@
+/**
+ * This file is our main entry point in the code.
+ */
+
+import React from 'react';
+import {MainAppCtl} from './MainApp/MainAppCtl';
+
+export const Main = () => {
+  return <MainAppCtl />;
+};
